feat: move the add folders command into a separate file under the `commands` directory
